backend/requirements.txt
Django==4.2.29
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
django-modeltranslation==0.19.11
django-mptt==0.16.0
django-storages==1.14.4
djangorestframework-simplejwt==5.4.0
dj-database-url==2.3.0
psycopg2-binary==2.9.10
Pillow==11.1.0
celery==5.4.0
redis==5.2.1
stripe==11.4.1
gunicorn==23.0.0
whitenoise==6.8.2
python-decouple==3.8
boto3==1.36.2
django-extensions==3.2.3
